// types/response.type.ts
export interface ResponseType<T> {
  data: T | null;
  message?: string;
  status: ResponseStatus;
}

enum ResponseStatus {
  SUCCESS = 200,
  CREATED = 201,
  ACCEPTED = 202,
  NO_CONTENT = 204,
  CREATE_ERROR = 207,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  INTERNAL_SERVER_ERROR = 500,
}

export { ResponseStatus };